Continue to split out Data into logical chunks.
Working towards #19.
