+++ boris-build/src/Boris/Build.hs
@@ -7,6 +7,10 @@ module Boris.Build (
   ) where
 
 import           Boris.Core.Data
+import           Boris.Core.Data.Build
+import           Boris.Core.Data.Configuration
+import           Boris.Core.Data.Repository
+import           Boris.Core.Data.Workspace
 
 import           Control.Monad.IO.Class (liftIO)
