+++ boris-client/src/Boris/Client/Build.hs
@@ -17,6 +17,8 @@ module Boris.Client.Build (
 import           Boris.Client.Http (BorisHttpClientError (..))
 import qualified Boris.Client.Http as H
 import           Boris.Core.Data
+import           Boris.Core.Data.Build
+import           Boris.Core.Data.Project
 import           Boris.Representation.ApiV1
 
 import           Data.Aeson (object, (.=))
