+++ boris-client/src/Boris/Client/Discover.hs
@@ -6,8 +6,10 @@ module Boris.Client.Discover (
 
 import           Boris.Client.Http (BorisHttpClientError (..))
 import qualified Boris.Client.Http as H
-import           Boris.Core.Data
+import           Boris.Core.Data.Build
 import           Boris.Core.Data.Instance
+import           Boris.Core.Data.Project
+
 import           Boris.Representation.ApiV1
 
 import           P
