+++ boris-client/src/Boris/Client/Log.hs
@@ -8,6 +8,8 @@ import           Boris.Client.Http (BorisHttpClientError (..))
 import qualified Boris.Client.Http as H
 import           Boris.Core.Data
 import           Boris.Core.Data.Log
+import           Boris.Core.Data.Build
+import           Boris.Core.Data.Project
 import           Boris.Representation.ApiV1
 
 import           P
