+++ boris-client/src/Boris/Client/Project.hs
@@ -9,6 +9,8 @@ module Boris.Client.Project (
 import           Boris.Client.Http (BorisHttpClientError (..))
 import qualified Boris.Client.Http as H
 import           Boris.Core.Data
+import           Boris.Core.Data.Build
+import           Boris.Core.Data.Project
 import           Boris.Representation.ApiV1
 
 import           P
