+++ boris-client/src/Boris/Client/Scoreboard.hs
@@ -7,6 +7,8 @@ module Boris.Client.Scoreboard (
 import           Boris.Client.Http (BorisHttpClientError (..))
 import qualified Boris.Client.Http as H
 import           Boris.Core.Data
+import           Boris.Core.Data.Build
+import           Boris.Core.Data.Project
 
 import           Data.Aeson (FromJSON (..), withObject, (.:), (.:?))
 import           Data.Aeson.Types (Parser, Value)
