+++ boris-core/boris-core.cabal
@@ -36,9 +36,12 @@ library
   exposed-modules:
                     Boris.Core.Data
                     Boris.Core.Data.Agent
+                    Boris.Core.Data.Build
                     Boris.Core.Data.Configuration
                     Boris.Core.Data.Instance
                     Boris.Core.Data.Log
+                    Boris.Core.Data.Project
+                    Boris.Core.Data.Repository
                     Boris.Core.Data.Time
                     Boris.Core.Data.Workspace
