+++ boris-core/src/Boris/Core/Data.hs
@@ -1,29 +1,7 @@
-{-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE NoImplicitPrelude #-}
 {-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE PackageImports #-}
 module Boris.Core.Data (
-    Environment (..)
-  , Source (..)
-  , sourceFromInt
-  , sourceToInt
-  , Owner (..)
-  , OwnerId (..)
-  , OwnerName (..)
-  , OwnerType (..)
-  , ownerTypeFromInt
-  , ownerTypeToInt
-  , Definition (..)
-  , ProjectId (..)
-  , Project (..)
-  , Build (..)
-  , BuildId (..)
-  , Repository (..)
-  , LocalRepository (..)
-  , Commit (..)
-  , Ref (..)
-  , Pattern (..)
-  , Registration (..)
+    Registration (..)
   , BuildResult (..)
   , Acknowledge (..)
   , BuildCancelled (..)
@@ -37,156 +15,17 @@ module Boris.Core.Data (
   , parseBuildResult
   , renderRegistration
   , parseRegistration
-  , sortBuildIds
-  , newBuild
   ) where
 
-import qualified Data.List as L
-import qualified Data.Text as T
-import           Data.Time (UTCTime, formatTime, defaultTimeLocale)
+import           Boris.Core.Data.Build
+import           Boris.Core.Data.Project
+import           Boris.Core.Data.Repository
 
-import qualified Data.Map.Strict as M
+import qualified Data.Text as T
+import           Data.Time (UTCTime)
 
 import           P
 
-import           System.FilePath (FilePath, (</>))
-import qualified "Glob" System.FilePath.Glob as G
-
-newtype Environment =
-  Environment {
-      renderEnvironment :: Text
-    } deriving (Eq, Show)
-
-data Source =
-    GitHubSource
-  | BorisSource
-    deriving (Eq, Ord, Show, Enum, Bounded)
-
-sourceToInt :: Source -> Int64
-sourceToInt s =
-  case s of
-    GitHubSource ->
-      0
-    BorisSource ->
-      1
-
-sourceFromInt :: Int64 -> Maybe Source
-sourceFromInt n =
-  case n of
-    0 ->
-      Just GitHubSource
-    1 ->
-      Just BorisSource
-    _ ->
-      Nothing
-
-newtype OwnerId =
-  OwnerId {
-      getOwnerId :: Int64
-    } deriving (Eq, Ord, Show)
-
-newtype OwnerName =
-  OwnerName {
-      getOwnerName :: Int64
-    } deriving (Eq, Ord, Show)
-
--- FIX this isn't right anymore, need to update to reflect sketch
-data OwnerType =
-    GitHubOwner
-  | BorisUser
-  | BorisSystem
-    deriving (Eq, Ord, Show, Enum, Bounded)
-
-ownerTypeToInt :: OwnerType -> Int64
-ownerTypeToInt o =
-  case o of
-    GitHubOwner ->
-      0
-    BorisUser ->
-      1
-    BorisSystem ->
-      2
-
-ownerTypeFromInt :: Int64 -> Maybe OwnerType
-ownerTypeFromInt n =
-  case n of
-    0 ->
-      Just GitHubOwner
-    1 ->
-      Just BorisUser
-    2 ->
-      Just BorisSystem
-    _ ->
-      Nothing
-
-data Owner =
-  Owner {
-      ownerId :: OwnerId
-    , ownerName :: OwnerName
-    , ownerType :: OwnerType
-    } deriving (Eq, Ord, Show)
-
-data Definition =
-  Definition {
-      definitionId :: ProjectId
-    , definitionSource :: Source
-    , definitionOwner :: Owner
-    , definitionProject :: Project
-    , definitionRepository :: Repository
-    } deriving (Eq, Ord, Show)
-
-newtype Project =
-  Project {
-      renderProject :: Text
-    } deriving (Eq, Show, Ord)
-
-newtype ProjectId =
-  ProjectId {
-      getProjectId :: Int64
-    } deriving (Eq, Show, Ord)
-
-newtype Build =
-  Build {
-      renderBuild :: Text
-    } deriving (Eq, Show, Ord)
-
-newtype BuildId =
-  BuildId {
-      renderBuildId :: Text
-    } deriving (Eq, Show)
-
-
-instance Ord BuildId where
-  compare b1 b2 =
-    let
-      asInt = (readMaybe :: [Char] -> Maybe Int) . T.unpack . renderBuildId
-    in
-      asInt b1 `compare` asInt b2
-
-newtype Repository =
-  Repository {
-      renderRepository :: Text
-    } deriving (Eq, Show, Ord)
-
-newtype LocalRepository =
-  LocalRepository {
-      renderLocalRepository :: Text
-    } deriving (Eq, Show, Ord)
-
-newtype Ref =
-  Ref {
-      renderRef :: Text
-    } deriving (Eq, Show, Ord)
-
-newtype Pattern =
-  Pattern {
-      renderPattern :: Text
-    } deriving (Eq, Show, Ord)
-
-newtype Commit =
-  Commit {
-      renderCommit :: Text
-    } deriving (Eq, Show, Ord)
 
 data Registration =
   Registration {
@@ -233,15 +72,6 @@ parseRegistration t =
       Just $ Registration (Project p) (Repository r)
     _ ->
       Nothing
-
-sortBuildIds :: [BuildId] -> [BuildId]
-sortBuildIds =
-  L.reverse . L.sort
-
-newBuild :: Text -> Maybe Build
-newBuild b =
-  emptyOrValue (T.isInfixOf "/" b) $
-    Build b
 
 
 data BuildCancelled =
