+++ boris-core/src/Boris/Core/Serial/Command.hs
@@ -6,7 +6,7 @@ module Boris.Core.Serial.Command (
   , renderBorisConfigError
   ) where
 
-import           Boris.Core.Data
+import           Boris.Core.Data.Build
 import           Boris.Core.Data.Configuration
 
 import           Control.Lens ((^?))
