+++ boris-core/test/Test/Boris/Core/Arbitrary.hs
@@ -8,8 +8,9 @@ import qualified Data.Text as T
 import           Disorder.Corpus
 
 import           Boris.Core.Data
+import           Boris.Core.Data.Build
 import           Boris.Core.Data.Configuration
-import           Boris.Core.Data.Log
+import           Boris.Core.Data.Project
 
 import           P
