+++ boris-http/boris-http.cabal
@@ -16,8 +16,6 @@ library
                     , aeson                           >= 0.8        && < 1.3
                     , ambiata-bmx
                     , ambiata-boris-core
-                    , ambiata-boris-queue
-                    , ambiata-boris-service
                     , ambiata-boris-representation
                     , ambiata-p
                     , ambiata-snooze
@@ -101,7 +99,6 @@ executable boris-http
                       base
                     , ambiata-boris-core
                     , ambiata-boris-http
-                    , ambiata-boris-queue
                     , ambiata-p
                     , ambiata-twine
                     , ambiata-x-eithert
@@ -133,7 +130,6 @@ test-suite test
                     , ambiata-boris-client
                     , ambiata-boris-http
                     , ambiata-boris-representation
-                    , ambiata-boris-queue
                     , ambiata-disorder-core
                     , ambiata-disorder-corpus
                     , ambiata-p
@@ -170,7 +166,6 @@ test-suite test-io
                     , ambiata-boris-core
                     , ambiata-boris-core-test
                     , ambiata-boris-http
-                    , ambiata-boris-queue
                     , ambiata-disorder-core
                     , ambiata-disorder-corpus
                     , ambiata-p
