+++ boris-http/src/Boris/Http/Api/Build.hs
@@ -24,6 +24,8 @@ import qualified Data.Text as Text
 import qualified Data.Time as Time
 
 import           Boris.Core.Data
+import           Boris.Core.Data.Build
+import           Boris.Core.Data.Project
 import           Boris.Core.Data.Log
 import qualified Boris.Http.Api.Project as Project
 import           Boris.Http.Data
