+++ boris-http/src/Boris/Http/Api/Discover.hs
@@ -10,8 +10,9 @@ module Boris.Http.Api.Discover (
 
 import qualified Data.List as List
 
-import           Boris.Core.Data
+import           Boris.Core.Data.Build
 import           Boris.Core.Data.Instance
+import           Boris.Core.Data.Project
 import qualified Boris.Http.Db.Query as Query
 
 import           P
