+++ boris-http/src/Boris/Http/Api/Project.hs
@@ -9,6 +9,9 @@ module Boris.Http.Api.Project (
   ) where
 
 import           Boris.Core.Data
+import           Boris.Core.Data.Build
+import           Boris.Core.Data.Project
+import           Boris.Core.Data.Repository
 import           Boris.Http.Data
 import qualified Boris.Http.Db.Query as Query
